Move items from src/app-controllers/common to where they belong
