+++ src/components/app-controllers/DualAppController/DualAppController.js
@@ -42,17 +42,20 @@ import {
 } from 'pcic-react-components';
 import { getMetadata } from '../../../data-services/ce-backend';
 import {
-  ensemble_name, filterOutMonthlyMym,
+  ensemble_name,
+} from '../common';
+import { setNamedState } from '../../../core/react-component-utils';
+import withAsyncMetadata from '../../../HOCs/withAsyncMetadata';
+import {
   findModelNamed, findScenarioIncluding, findVariableMatching,
   representativeValue, constraintsFor, filterMetaBy,
-  setState, withMetadata
-} from '../common';
+} from '../../../core/selectors';
 
 
 class DualAppControllerDisplay extends React.Component {
   // This is a pure (state-free), controlled component that renders the
   // entire content of DualAppController, including the controls.
-  // It is wrapped by `withMetadata` to inject the asynchronously fetched
+  // It is wrapped by `withAsyncMetadata` to inject the asynchronously fetched
   // metadata that it needs.
 
   static propTypes = {
@@ -224,7 +227,7 @@ class DualAppControllerDisplay extends React.Component {
 
 
 // Inject asynchronously fetched metadata into controlled component.
-const WmdDualAppControllerDisplay = withMetadata(DualAppControllerDisplay);
+const WmdDualAppControllerDisplay = withAsyncMetadata(DualAppControllerDisplay);
 
 
 export default class DualAppController extends React.Component {
@@ -239,11 +242,11 @@ export default class DualAppController extends React.Component {
   };
 
   // TODO: https://github.com/pacificclimate/climate-explorer-frontend/issues/122
-  handleChangeArea = setState(this, 'area');
-  handleChangeModel = setState(this, 'model');
-  handleChangeScenario = setState(this, 'scenario');
-  handleChangeVariable = setState(this, 'variable');
-  handleChangeComparand = setState(this, 'comparand');
+  handleChangeArea = setNamedState(this, 'area');
+  handleChangeModel = setNamedState(this, 'model');
+  handleChangeScenario = setNamedState(this, 'scenario');
+  handleChangeVariable = setNamedState(this, 'variable');
+  handleChangeComparand = setNamedState(this, 'comparand');
 
   render() {
     return (
